rework app component to properly display content on the page
