+++ src/index.css
@@ -4,6 +4,10 @@
   padding: 0;
 }
 
+html {
+  overflow: hidden;
+}
+
 button {
   outline: none;
   border: none;
